hello: Initial commit of IRCdns
This is a small hack to build a cheap, reliable dynamic dns service
using existing IRC servers as the relay. At the moment, there is just
the publish service. A human is required to read the ip address
manually. In the future we might be able to get these ip addresses
automatically with a subscriber client, nsswitch magic and more code. We
might consider adding public key encryption so that our published
messages are slightly more secure, and more!

Happy Hacking!