client: Add an nss client
This adds a libc name switch service client so that lookups can be
automatic! The performance isn't great because connecting to IRC is
slow, but it is well within what we require for our use-case. To improve
performance we can utilize caching as well as an on-demand or
continuously running systemd service to keep a local cache.

All of this was possible once I knew how to build the .so file in
golang, and that is all thanks to Leo Antunes (@costela) who worked out
the C parts previously. Thanks!